+++ plugin/src/main/scala/com/stratio/cassandra/lucene/IndexQueryHandler.scala
@@ -19,7 +19,7 @@ import java.nio.ByteBuffer
 import java.{util => java}
 
 import com.stratio.cassandra.lucene.IndexQueryHandler._
-import com.stratio.cassandra.lucene.util.TimeCounter
+import com.stratio.cassandra.lucene.util.{Logging, TimeCounter}
 import org.apache.cassandra.cql3._
 import org.apache.cassandra.cql3.statements.RequestValidations.checkNotNull
 import org.apache.cassandra.cql3.statements.{BatchStatement, IndexTarget, ParsedStatement, SelectStatement}
@@ -32,17 +32,16 @@ import org.apache.cassandra.service.{LuceneStorageProxy, QueryState}
 import org.apache.cassandra.transport.messages.ResultMessage
 import org.apache.cassandra.transport.messages.ResultMessage.{Prepared, Rows}
 import org.apache.cassandra.utils.{FBUtilities, MD5Digest}
-import org.slf4j.LoggerFactory
 
-import scala.collection.mutable
 import scala.collection.JavaConverters._
+import scala.collection.mutable
 
 
 /** [[QueryHandler]] to be used with Lucene searches.
   *
   * @author Andres de la Pena `adelapena@stratio.com`
   */
-class IndexQueryHandler extends QueryHandler {
+class IndexQueryHandler extends QueryHandler with Logging {
 
   type Payload = java.Map[String, ByteBuffer]
 
@@ -134,14 +133,14 @@ class IndexQueryHandler extends QueryHandler {
     val expressions = select.getRowFilter(options).getExpressions
     val cfs = Keyspace.open(select.keyspace).getColumnFamilyStore(select.columnFamily)
     val indexes = cfs.indexManager.listIndexes.asScala.collect { case index: Index => index }
-    expressions.asScala.foreach {
+    expressions.forEach {
       case expression: CustomExpression =>
         val clazz = expression.getTargetIndex.options.get(IndexTarget.CUSTOM_INDEX_OPTION_NAME)
         if (clazz == classOf[Index].getCanonicalName) {
           val index = cfs.indexManager.getIndex(expression.getTargetIndex).asInstanceOf[Index]
           map += expression -> index
         }
-      case expr =>
+      case expr: Expression =>
         indexes.filter(_.supportsExpression(expr.column, expr.operator)).foreach(map.put(expr, _))
     }
     map.toMap
@@ -223,9 +222,8 @@ class IndexQueryHandler extends QueryHandler {
   }
 }
 
+/** Companion object for [[IndexQueryHandler]]. */
 object IndexQueryHandler {
-
-  val logger = LoggerFactory.getLogger(classOf[IndexQueryHandler])
 
   val getPageSize = classOf[SelectStatement].getDeclaredMethod("getPageSize", classOf[QueryOptions])
   getPageSize.setAccessible(true)