+++ plugin/src/main/scala/com/stratio/cassandra/lucene/mapping/ClusteringMapper.scala
@@ -61,13 +61,13 @@ class ClusteringMapper(metadata: CFMetaData) {
     * @return the columns
     */
   def columns(clustering: Clustering): Columns = {
-    clusteringColumns.foldLeft(new Columns)(
+    (Columns() /: clusteringColumns) (
       (columns, columnDefinition) => {
         val name = columnDefinition.name.toString
         val position = columnDefinition.position
         val value = clustering.get(position)
         val valueType = columnDefinition.cellValueType
-        columns.add(Column.apply(name).withValue(ColumnsMapper.compose(value, valueType)))
+        columns.add(Column(name).withValue(ColumnsMapper.compose(value, valueType)))
       })
   }
 
@@ -78,7 +78,6 @@ class ClusteringMapper(metadata: CFMetaData) {
     * @return a indexable field
     */
   def indexableFields(key: DecoratedKey, clustering: Clustering): List[IndexableField] = {
-    // TODO: return Seq
 
     // Build stored field for clustering key retrieval
     val plainClustering = bytesRef(byteBuffer(clustering))
@@ -98,7 +97,7 @@ class ClusteringMapper(metadata: CFMetaData) {
     * @return a byte buffer representing `clustering`
     */
   def byteBuffer(clustering: Clustering): ByteBuffer = {
-    clustering.getRawValues.foldLeft(clusteringType.builder)(_ add _).build()
+    (clusteringType.builder /: clustering.getRawValues) (_ add _) build()
   }
 
   /** Returns the [[String]] human-readable representation of the specified [[ClusteringPrefix]].
@@ -168,14 +167,13 @@ class ClusteringMapper(metadata: CFMetaData) {
     * @return the Lucene query
     */
   def query(key: DecoratedKey, filter: ClusteringIndexSliceFilter): Query = {
-    filter.requestedSlices.asScala.foldLeft(new BooleanQuery.Builder)(
-      (builder, slice) => {
-        builder.add(query(key, slice), SHOULD)
-      }).build
+    (new BooleanQuery.Builder /: filter.requestedSlices.asScala) (
+      (builder, slice) => builder.add(query(key, slice), SHOULD)).build()
   }
 
 }
 
+/** Companion object for [[ClusteringMapper]]. */
 object ClusteringMapper {
 
   /** The Lucene field name. */
@@ -249,23 +247,15 @@ object ClusteringMapper {
   * @param mapper the primary key mapper to be used
   */
 class ClusteringSort(mapper: ClusteringMapper) extends SortField(
-  FIELD_NAME, new FieldComparatorSource {
-    override def newComparator(
-        field: String,
-        hits: Int,
-        sortPos: Int,
-        reversed: Boolean): FieldComparator[_] = {
-      new TermValComparator(hits, field, false) {
-        override def compareValues(t1: BytesRef, t2: BytesRef): Int = {
-          val comp = compareUnsigned(t1.bytes, 0, PREFIX_SIZE, t2.bytes, 0, PREFIX_SIZE)
-          if (comp != 0) return comp
-          val bb1 = ByteBuffer.wrap(t1.bytes, PREFIX_SIZE, t1.length - PREFIX_SIZE)
-          val bb2 = ByteBuffer.wrap(t2.bytes, PREFIX_SIZE, t2.length - PREFIX_SIZE)
-          val clustering1 = mapper.clustering(bb1)
-          val clustering2 = mapper.clustering(bb2)
-          mapper.comparator.compare(clustering1, clustering2)
-        }
-      }
+  FIELD_NAME, (field, hits, sortPos, reversed) => new TermValComparator(hits, field, false) {
+    override def compareValues(t1: BytesRef, t2: BytesRef): Int = {
+      val comp = compareUnsigned(t1.bytes, 0, PREFIX_SIZE, t2.bytes, 0, PREFIX_SIZE)
+      if (comp != 0) return comp
+      val bb1 = ByteBuffer.wrap(t1.bytes, PREFIX_SIZE, t1.length - PREFIX_SIZE)
+      val bb2 = ByteBuffer.wrap(t2.bytes, PREFIX_SIZE, t2.length - PREFIX_SIZE)
+      val clustering1 = mapper.clustering(bb1)
+      val clustering2 = mapper.clustering(bb2)
+      mapper.comparator.compare(clustering1, clustering2)
     }
   }) {