+++ plugin/src/main/scala/com/stratio/cassandra/lucene/util/Logging.scala
@@ -13,29 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.stratio.cassandra.lucene.util
 
-import org.apache.cassandra.tracing.{Tracing => Tracer}
+import com.typesafe.scalalogging.{Logger, StrictLogging}
 
-/** Wrapper for [[Tracer]] avoiding test environment failures.
+/** Trait including a [[Logger]] instance named `logger`.
   *
   * @author Andres de la Pena `adelapena@stratio.com`
   */
-final class Tracer extends Logging {
-
-  /** If Cassandra tracing is enabled. */
-  lazy val canTrace: Boolean = try {
-    Tracer.isTracing
-    true
-  } catch {
-    case e: Error =>
-      logger.warn(s"Unable to trace: ${e.getMessage}", e)
-      false
-  }
-
-  /** Traces the specified string message.
-    *
-    * @param message the message to be traced
-    */
-  def trace(message: => String) = if (canTrace && Tracer.isTracing) Tracer.trace(message)
-}
+trait Logging extends StrictLogging {}