+++ plugin/src/main/scala/com/stratio/cassandra/lucene/util/TaskQueue.scala
@@ -16,14 +16,13 @@
 package com.stratio.cassandra.lucene.util
 
 import java.io.Closeable
-import java.util.concurrent._
 import java.util.concurrent.TimeUnit.DAYS
+import java.util.concurrent._
 import java.util.concurrent.locks.ReentrantReadWriteLock
 
 import com.stratio.cassandra.lucene.IndexException
-import com.stratio.cassandra.lucene.util.JavaConversions.asJavaCallable
-import com.stratio.cassandra.lucene.util.TaskQueue._
-import org.slf4j.LoggerFactory
+
+//import com.stratio.cassandra.lucene.util.JavaConversions.asJavaCallable
 
 import scala.concurrent.ExecutionException
 
@@ -33,7 +32,7 @@ import scala.concurrent.ExecutionException
   *
   * @author Andres de la Pena `adelapena@stratio.com`
   */
-sealed trait TaskQueue extends Closeable {
+sealed trait TaskQueue extends Closeable with Logging {
 
   /** Submits a non value-returning task for asynchronous execution.
     *
@@ -79,16 +78,14 @@ private class TaskQueueAsync(numThreads: Int, queuesSize: Int) extends TaskQueue
   private val lock = new ReentrantReadWriteLock(true)
   private val pools = (1 to numThreads)
     .map(index => new ArrayBlockingQueue[Runnable](queuesSize, true))
-    .map(queue => new ThreadPoolExecutor(1, 1, 1, DAYS, queue, new RejectedExecutionHandler() {
-      override def rejectedExecution(task: Runnable, executor: ThreadPoolExecutor) =
-        if (!executor.isShutdown) executor.getQueue.put(task)
-    }))
+    .map(queue => new ThreadPoolExecutor(1, 1, 1, DAYS, queue,
+      (task, executor) => if (!executor.isShutdown) executor.getQueue.put(task)))
 
   /** @inheritdoc */
   override def submitAsynchronous[A](id: AnyRef, task: () => A): Unit = {
     lock.readLock.lock()
     try {
-      pools(Math.abs(id.hashCode % numThreads)).submit(task)
+      pools(Math.abs(id.hashCode % numThreads)).submit(() => task.apply())
     } catch {
       case e: Exception =>
         logger.error("Task queue asynchronous submission failed", e)
@@ -98,9 +95,9 @@ private class TaskQueueAsync(numThreads: Int, queuesSize: Int) extends TaskQueue
 
   /** @inheritdoc */
   override def submitSynchronous[A](task: () => A): A = {
-    lock.writeLock().lock()
+    lock.writeLock.lock()
     try {
-      pools.map(_.submit(() => {})).foreach(_.get()) // Wait for queued tasks completion
+      pools.map(_.submit(() => None)).map(_.get()) // Wait for queued tasks completion
       task.apply // Run synchronous task
     } catch {
       case e: InterruptedException =>
@@ -124,9 +121,8 @@ private class TaskQueueAsync(numThreads: Int, queuesSize: Int) extends TaskQueue
 
 }
 
+/** Companion object for [[TaskQueue]]. */
 object TaskQueue {
-
-  val logger = LoggerFactory.getLogger(classOf[TaskQueue])
 
   /** Returns a new [[TaskQueue]].
     *
@@ -135,7 +131,7 @@ object TaskQueue {
     * @return a new task queue
     */
   def build(numThreads: Int, queuesSize: Int): TaskQueue = {
-    if (numThreads > 0) new TaskQueueAsync(numThreads, queuesSize) else new TaskQueueSync()
+    if (numThreads > 0) new TaskQueueAsync(numThreads, queuesSize) else new TaskQueueSync
   }
 
 }